Added The Unlicense with original author's statement
The original author's license says, "This software is free. It is distributed with the source code and you are encouraged to do anything you want with it. If you do anything interesting, I'd be happy to hear about it."

That seems closest in intent to The Unlicense, which applies to any code modifications under Bill's Rebuild.